Configura acesso ao authorization server
