+++ pom.xml
@@ -19,6 +19,7 @@
         <jasperreports.version>6.16.0</jasperreports.version>
         <aws-java-sdk-s3.version>1.11.671</aws-java-sdk-s3.version>
         <springfox.version>2.9.2</springfox.version>
+        <oauth2-oidc-sdk.version>9.1</oauth2-oidc-sdk.version>
     </properties>
 
     <dependencies>
@@ -153,6 +154,15 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-security</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-oauth2-resource-server</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.nimbusds</groupId>
+            <artifactId>oauth2-oidc-sdk</artifactId>
+            <version>${oauth2-oidc-sdk.version}</version>
         </dependency>
     </dependencies>
     <build>
