+++ src/main/resources/application.properties
@@ -51,3 +51,8 @@ sentry.logging.minimum-breadcrumb-level=debug
 
 spring.security.user.name=algafood
 spring.security.user.password=245ebbd5
+
+# Configura\u00E7\u00F5es do resource server
+spring.security.oauth2.resourceserver.opaquetoken.introspection-uri=http://localhost:8081/oauth/check_token
+spring.security.oauth2.resourceserver.opaquetoken.client-id=refresh-token
+spring.security.oauth2.resourceserver.opaquetoken.client-secret=web123
